plotting functions moved to the new abstraction
