+++ crates/subspace-farmer/src/lib.rs
@@ -22,6 +22,7 @@ pub(crate) mod farming;
 pub(crate) mod identity;
 pub(crate) mod object_mappings;
 pub(crate) mod plot;
+pub(crate) mod plotting;
 pub(crate) mod rpc;
 
 pub use commands::farm;
