+++ crates/subspace-farmer/src/main.rs
@@ -22,9 +22,9 @@ mod farming;
 mod identity;
 mod object_mappings;
 mod plot;
+mod plotting;
 mod rpc;
 mod utils;
-
 use anyhow::Result;
 use clap::{Parser, ValueHint};
 use env_logger::Env;
